Added params and test file init

+++ src/main/java/com/bezkoder/spring/login/controllers/UserController.java
@@ -3,6 +3,8 @@ package com.bezkoder.spring.login.controllers;
 import com.bezkoder.spring.login.models.User;
 import com.bezkoder.spring.login.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.support.NullValue;
+import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -19,6 +21,11 @@ public class UserController {
   @Autowired
   UserService userService;
 
+  @GetMapping("/all")
+  public String allAccess() {
+    return "Public Content.";
+  }
+
   @GetMapping("/{id}")
   public ResponseEntity<?> getUser(@PathVariable Long id) {
     Optional<User> user = userService.findById(id);
@@ -27,33 +34,45 @@ public class UserController {
     }
     return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not found");
   }
-  @GetMapping("/all")
-  public String allAccess() {
-    return "Public Content.";
-  }
 
   @GetMapping("/candidates")
   @PreAuthorize("hasRole('ADMIN')")
-  public List<User> getCandidates(){
-      return userService.findCandidates();
+  public ResponseEntity<?> getCandidates(){
+      List<User> candidates = userService.findCandidates();
+    if(!candidates.isEmpty()){
+      return ResponseEntity.status(HttpStatus.FOUND).body(candidates);
+    }
+    return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not found");
   }
 
   @GetMapping("/companies")
   @PreAuthorize("hasRole('ADMIN')")
-  public List<User> getCompanies(){
-      return userService.findCompanies();
+  public ResponseEntity<?> getCompanies(){
+    List<User> companies = userService.findCompanies();
+    if(!companies.isEmpty()){
+      return ResponseEntity.status(HttpStatus.FOUND).body(companies);
+    }
+    return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not found");
   }
 
   @PutMapping("/{id}")
   @PreAuthorize("hasRole('ADMIN')")
-  public User editUser(@RequestBody User user, @PathVariable Long id){
-      return userService.updateUser(id, user);
+  public ResponseEntity<?> editUser(@RequestBody User user, @PathVariable Long id){
+      if(userService.findById(id).isPresent()){
+        userService.updateUser(id, user);
+        return ResponseEntity.status(HttpStatus.OK).body("Successfully changed");
+      }
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not found");
   }
 
   @DeleteMapping("/{id}")
   @PreAuthorize("hasRole('ADMIN')")
-  public void deleteUser(@PathVariable Long id){
-     userService.deleteUser(id);
+  public ResponseEntity<?> deleteUser(@PathVariable Long id){
+    if(userService.findById(id).isPresent()){
+      userService.deleteUser(id);
+      return ResponseEntity.status(HttpStatus.OK).body("Successfully deleted");
+    }
+    return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not found");
   }
 
 }
